# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      err_code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  TypeAuth:
    properties:
      auth:
        additionalProperties:
          $ref: '#/definitions/TypeAuth'
        type: object
      refreshTime:
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyLocalResponse:
    properties:
      code:
        example: 200
        type: integer
      val:
        example: Cluster Config insert Success
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyRequest:
    properties:
      action:
        description: 'action: insert/remove/reset/check'
        example: insert
        type: string
      ccvm:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterCCVMConfig'
        description: ccvm config (ip only)
      exclude_hostname:
        description: exclude hostname (hci only)
        example: ablecube31-2
        type: string
      external_timeserver:
        description: external timeserver
        example: time.google.com
        type: string
      hosts:
        description: hosts list (preferred)
        items:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterHost'
        type: array
      iscsi_storage:
        description: iscsi storage usage (true/false)
        example: "false"
        type: string
      mngtNic:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterMngtNicConfig'
        description: management NIC config
      new_hostname:
        description: 'new hostname (option=add: exclude scvm fanout)'
        example: ablecube12-3
        type: string
      option:
        description: option (e.g. add/hostOnly/withScvm/all/local)
        example: add
        type: string
      pcs_cluster_list:
        description: pcs cluster node IP list
        example:
        - 10.10.31.1
        - 10.10.31.2
        - 10.10.31.3
        items:
          type: string
        type: array
      remove_hostname:
        description: remove hostname
        example: ablecube31-3
        type: string
      type:
        description: cluster type (e.g. ablestack-vm, ablestack-hci)
        example: ablestack-vm
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterCCVMConfig:
    properties:
      ip:
        description: ccvm IP
        example: 10.10.31.10
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterHost:
    properties:
      ablecube:
        example: 10.10.31.1
        type: string
      ablecubePn:
        example: 100.100.31.1
        type: string
      hostname:
        example: ablecube31-1
        type: string
      index:
        example: "1"
        type: string
      scvm:
        example: 100.100.31.11
        type: string
      scvmCn:
        example: 100.200.31.11
        type: string
      scvmMngt:
        example: 10.10.31.11
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterMngtNicConfig:
    properties:
      cidr:
        description: management NIC CIDR
        example: "16"
        type: string
      dns:
        description: management NIC DNS
        example: 8.8.8.8
        type: string
      gw:
        description: management NIC gateway
        example: 10.10.0.1
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_service_authservice.TokenClaim:
    properties:
      exp:
        type: integer
      iat:
        type: integer
      role:
        type: string
      sub:
        type: string
      use:
        type: string
    type: object
  internal_handler_auth.AuthApplyRequest:
    properties:
      token:
        type: string
    type: object
  internal_handler_auth.AuthSyncRequest:
    properties:
      option:
        enum:
        - host
        - scvm
        - ccvm
        - all
        example: all
        type: string
    type: object
  internal_handler_auth.AuthSyncResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        type: string
      results:
        items:
          $ref: '#/definitions/internal_handler_auth.AuthSyncResult'
        type: array
      val:
        additionalProperties: {}
        type: object
    type: object
  internal_handler_auth.AuthSyncResult:
    properties:
      code:
        type: integer
      hostname:
        type: string
      message:
        type: string
      role:
        type: string
      target:
        type: string
    type: object
  internal_handler_auth.LoginRequest:
    properties:
      id:
        example: root
        type: string
      password:
        example: password
        type: string
    type: object
  internal_handler_auth.LoginResponse:
    properties:
      access_token:
        type: string
      authorization:
        example: Bearer eyJ...
        type: string
      code:
        example: 200
        type: integer
      expires_in:
        example: 3600
        type: integer
      token_type:
        example: Bearer
        type: string
    type: object
  internal_handler_auth.MeResponse:
    properties:
      code:
        example: 200
        type: integer
      val:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_service_authservice.TokenClaim'
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
  /auth/apply:
    post:
      consumes:
      - application/json
      description: 내부 호출로 전달받은 API 인증 값을 현재 호스트에 적용합니다.
      parameters:
      - description: auth apply request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_auth.AuthApplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      summary: Apply Auth
      tags:
      - Auth
  /auth/login:
    post:
      security: []
      consumes:
      - application/json
      description: 웹 UI/API용 access token을 발급합니다.
      parameters:
      - description: login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_auth.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - Auth
  /auth/me:
    get:
      description: Bearer token을 검증하고 현재 token claim을 반환합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_auth.MeResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Current User
      tags:
      - Auth
  /auth/sync:
    post:
      consumes:
      - application/json
      description: 현재 호스트의 API 인증 값을 선택한 대상(host/scvm/ccvm/all) API 서버에 동기화합니다. host는
        hosts[].ablecube, scvm은 hosts[].scvm, ccvm은 ccvm.ip를 사용합니다. ablestack-vm/ablestack-standalone의
        all은 host,ccvm만 포함합니다.
      parameters:
      - description: sync request
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler_auth.AuthSyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_auth.AuthSyncResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Sync Auth
      tags:
      - Auth
...
  /cube/cluster/apply-local:
    post:
      consumes:
      - application/json
      description: 로컬 노드에서만 cluster_config CLI를 실행합니다. insert 적용 시 시간 서버 설정도 함께 적용합니다.
      parameters:
      - description: apply request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyLocalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Apply Cluster Config (Local)
      tags:
      - CUBE - Cluster
...
  /glue/auth:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: GLUE의 인증키를 보여줍니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TypeAuth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show Auth of GLUE
      tags:
      - API
      - Glue
      - GLUE
  /glue/auths:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: GLUE의 인증키 목록을 보여줍니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TypeAuth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show Auths of GLUE
      tags:
      - API
      - Glue
      - GLUE
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter the token with the Bearer prefix, e.g. Bearer eyJ...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"